# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Hospital patient management API",
        "description": "REST API with PHP for hospital patient management, which will allow hospital doctors to search for a patient, create new patients and add diagnoses to patients.",
        "contact": {
            "name": "Jorge Ivan Carrillo",
            "url": "https://www.linkedin.com/in/jorgecarrillog/"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/diagnoses": {
            "post": {
                "tags": [
                    "Diagnoses"
                ],
                "summary": "Create a new diagnosis",
                "description": "Registers a new diagnosis in the system.",
                "operationId": "createDiagnosis",
                "requestBody": {
                    "description": "Diagnosis data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Diagnosis name",
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "description": {
                                        "description": "Diagnosis description",
                                        "type": "string",
                                        "maxLength": 255
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "Example Diagnosis",
                                    "description": "This is an example diagnosis"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "string",
                                            "example": "Diagnóstico creado correctamente."
                                        },
                                        "diagnosis": {
                                            "$ref": "#/components/schemas/Diagnosis"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "419": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "name": [
                                                    "The name field is required."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Error al crear el diagnóstico."
                                        },
                                        "details": {
                                            "type": "string",
                                            "example": "Error details"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/diagnoses/{patient_id}/assign-diagnosis": {
            "post": {
                "tags": [
                    "Diagnoses"
                ],
                "summary": "Assign a diagnosis to a patient",
                "description": "Assigns a diagnosis to a specific patient in the system.",
                "operationId": "assignDiagnosisToPatient",
                "parameters": [
                    {
                        "name": "patient_id",
                        "in": "path",
                        "description": "ID of the patient",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Diagnosis data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "diagnosis_id",
                                    "diagnosis_date"
                                ],
                                "properties": {
                                    "diagnosis_id": {
                                        "description": "ID of the diagnosis",
                                        "type": "integer"
                                    },
                                    "observation": {
                                        "description": "Observation about the diagnosis",
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "diagnosis_date": {
                                        "description": "Diagnosis date",
                                        "type": "string",
                                        "format": "date"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "diagnosis_id": 1,
                                    "observation": "Observation text",
                                    "diagnosis_date": "2023-12-31"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "string",
                                            "example": "Diagnóstico asignado correctamente al paciente."
                                        },
                                        "patient": {
                                            "$ref": "#/components/schemas/Patient"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "419": {
                        "description": "Validation error or Patient not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "diagnosis_date": [
                                                    "The diagnosis date field is required."
                                                ]
                                            }
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "El paciente no fue encontrado."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Error al asignar el diagnóstico al paciente."
                                        },
                                        "details": {
                                            "type": "string",
                                            "example": "Error details"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/diagnoses/top-diagnoses-last-six-months": {
            "get": {
                "tags": [
                    "Diagnoses"
                ],
                "summary": "Get top diagnoses in the last six months",
                "description": "Retrieves the top five diagnoses that have been assigned the most in the last six months.",
                "operationId": "topDiagnosesLastSixMonths",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "top_diagnoses": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Diagnosis"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Error al obtener los diagnósticos más asignados."
                                        },
                                        "details": {
                                            "type": "string",
                                            "example": "Error details"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Diagnosis": {
                "title": "Diagnosis",
                "required": [
                    "name"
                ],
                "properties": {
                    "id": {
                        "description": "Unique identifier for the diagnosis",
                        "type": "integer",
                        "format": "int64",
                        "readOnly": true
                    },
                    "name": {
                        "description": "Diagnostic name",
                        "type": "string",
                        "maxLength": 255
                    },
                    "description": {
                        "description": "Description of the diagnosis",
                        "type": "string",
                        "maxLength": 255
                    }
                },
                "type": "object",
                "example": {
                    "name": "Diabetes",
                    "description": "Type 2 diabetes mellitus"
                }
            },
...
            "Patient": {
                "title": "Patient",
                "required": [
                    "document",
                    "first_name",
                    "last_name",
                    "birth_date",
                    "email",
                    "phone",
                    "genre"
                ],
                "properties": {
                    "id": {
                        "description": "Identificator in system",
                        "type": "integer",
                        "format": "int64",
                        "readOnly": true
                    },
                    "document": {
                        "description": "Identification document",
                        "type": "integer",
                        "format": "int64",
                        "maxLength": 20,
                        "uniqueItems": true
                    },
                    "first_name": {
                        "description": "Patient first name",
                        "type": "string",
                        "maxLength": 255
                    },
                    "last_name": {
                        "description": "Patient last name",
                        "type": "string",
                        "maxLength": 255
                    },
                    "birth_date": {
                        "description": "Patient birthday",
                        "type": "string",
                        "format": "date"
                    },
                    "email": {
                        "description": "Contact email",
                        "type": "string",
                        "format": "email",
                        "maxLength": 255,
                        "uniqueItems": true
                    },
                    "phone": {
                        "description": "Contact phone",
                        "type": "string",
                        "maxLength": 20
                    },
                    "genre": {
                        "description": "Patient genre (Male/Female)",
                        "type": "string",
                        "enum": [
                            "Male",
                            "Female"
                        ]
                    }
                },
                "type": "object",
                "example": {
                    "document": 9223372036854775807,
                    "first_name": "John",
                    "last_name": "Doe",
                    "birth_date": "1990-01-01",
                    "email": "johndoe@example.com",
                    "phone": 1234567890,
                    "genre": "Male"
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Bearer Token Authentication",
                "name": "Authorization",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}